Detection du datatype coté C++

+++ examples/mdspan-interop2/tests/matrix_product_test.rs
@@ -3,8 +3,8 @@ use::mdspan_interop2::{matrix_product};
 
 #[test] 
 fn matrix_product_test() {
-    let v: [f32; 12] = [0.0, 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0];
-    let s: [f32; 12] = [0.0, 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0, 10.0, 11.0];
+    let v: [u32; 12] = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11];
+    let s: [u32; 12] = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11];
     let arr1 = ArrayView::from_shape((2,2), &v).unwrap();
     let arr2 = ArrayView::from_shape((2,2), &s).unwrap();
